+++ pkg/output/file_json.go
@@ -16,18 +16,19 @@ type ejsonFileOutput struct {
 }
 
 type JsonResult struct {
-	Input            map[string]string `json:"input"`
-	Position         int               `json:"position"`
-	StatusCode       int64             `json:"status"`
-	ContentLength    int64             `json:"length"`
-	ContentWords     int64             `json:"words"`
-	ContentLines     int64             `json:"lines"`
-	ContentType      string            `json:"content-type"`
-	RedirectLocation string            `json:"redirectlocation"`
-	Duration         time.Duration     `json:"duration"`
-	ResultFile       string            `json:"resultfile"`
-	Url              string            `json:"url"`
-	Host             string            `json:"host"`
+	Input            map[string]string   `json:"input"`
+	Position         int                 `json:"position"`
+	StatusCode       int64               `json:"status"`
+	ContentLength    int64               `json:"length"`
+	ContentWords     int64               `json:"words"`
+	ContentLines     int64               `json:"lines"`
+	ContentType      string              `json:"content-type"`
+	RedirectLocation string              `json:"redirectlocation"`
+	ScraperData      map[string][]string `json:"scraper"`
+	Duration         time.Duration       `json:"duration"`
+	ResultFile       string              `json:"resultfile"`
+	Url              string              `json:"url"`
+	Host             string              `json:"host"`
 }
 
 type jsonFileOutput struct {
@@ -73,6 +74,7 @@ func writeJSON(filename string, config *ffuf.Config, res []ffuf.Result) error {
 			ContentLines:     r.ContentLines,
 			ContentType:      r.ContentType,
 			RedirectLocation: r.RedirectLocation,
+			ScraperData:      r.ScraperData,
 			Duration:         r.Duration,
 			ResultFile:       r.ResultFile,
 			Url:              r.Url,